+++ docker/erc20.Dockerfile
@@ -22,7 +22,7 @@ RUN source /opt/sgxsdk/environment && \
     source /root/.cargo/env && \
     export SGX_MODE=HW && \
     export RUSTFLAGS=-Ctarget-feature=+aes,+sse2,+sse4.1,+ssse3 && \
-    solc -o contract-build --bin --abi --optimize --overwrite contracts/Anonify.sol && \
+    solc -o contract-build --bin --abi --optimize --overwrite ethereum/contracts/Anonify.sol && \
     cd scripts && \
     make ENCLAVE_DIR=example/erc20/enclave ENCLAVE_PKG_NAME=erc20 CARGO_FLAGS=--release && \
     cd ../example/erc20/server && \
