+++ modules/anonify-eth-driver/src/eth/mod.rs
@@ -4,3 +4,4 @@ pub mod sender;
 
 pub use self::event_watcher::EventWatcher;
 pub use self::sender::EthSender;
+pub use self::connection::Web3Http;
