+++ scripts/build-server-in-docker.sh
@@ -29,7 +29,7 @@ make DEBUG=1 FEATURES=ERC20
 # enclave.signed.so is need to initialize enclave.
 rm -rf ../example/erc20/bin && cp -rf bin/ ../example/erc20/bin/ && cd ../
 
-solc -o contract-build --bin --abi --optimize --overwrite contracts/Anonify.sol
+solc -o contract-build --bin --abi --optimize --overwrite ethereum/contracts/Anonify.sol
 cd example/erc20/server
 
 echo "Build artifacts in debug mode."
