+++ scripts/run-server.sh
@@ -7,7 +7,7 @@ export PATH=~/.cargo/bin:$PATH
 export SGX_MODE=HW
 export RUSTFLAGS=-Ctarget-feature=+aes,+sse2,+sse4.1,+ssse3
 
-solc -o contract-build --bin --abi --optimize --overwrite contracts/Anonify.sol
+solc -o contract-build --bin --abi --optimize --overwrite ethereum/contracts/Anonify.sol
 
 dirpath=$(cd $(dirname $0) && pwd)
 cd "${dirpath}/.."
